Factor out acyclic package.
